Implement changing plot controls and overfreq plot
