+++ pychamber/deprecated/network_model.py
