+++ pychamber/plugins/__init__.py
@@ -1,4 +1,5 @@
 from .analyzer import AnalyzerPlugin
 from .base import PyChamberPlugin
 from .calibration import CalibrationPlugin
+from .plots import PlotsPlugin
 from .positioner import PositionerPlugin
