+++ pychamber/plugins/plots/__init__.py
@@ -1,0 +1,1 @@
+from .plots_container import PlotControls, PlotsPlugin
