+++ pychamber/plugins/plots/over_freq.py
@@ -1,9 +1,9 @@
-import numpy as np
 import skrf
 from PyQt5.QtCore import QStringListModel
 from PyQt5.QtWidgets import (
     QCheckBox,
     QComboBox,
+    QDoubleSpinBox,
     QHBoxLayout,
     QLabel,
     QPushButton,
@@ -14,65 +14,64 @@ from PyQt5.QtWidgets import (
 )
 
 from pychamber.logger import log
-from pychamber.ui import size_policy
-from pychamber.widgets import FrequencyLineEdit
 
-from ..freq_spin_box import FrequencySpinBox
-from .mpl_widget import MplPolarWidget, PlotLimits
+from .mpl_widget import MplRectWidget, PlotLimits
 from .pychamber_plot import PlotControls, PyChamberPlot
 
 
-class PolarPlot(PyChamberPlot):
+class OverFreqPlot(PyChamberPlot):
     def __init__(self, parent=None) -> None:
         super().__init__(parent)
 
-    def init_from_experiment(self, **kwargs) -> None:
-        freqs = kwargs.get('frequencies')
-        self.freq_spinbox.setRange(freqs.min(), freqs.max())
-        self.freq_spinbox.setSingleStep(freqs[1] - freqs[0])
+    def post_visible_setup(self) -> None:
+        self.plot.set_xlabel("Frequency [Hz]")
+        self.plot.set_ylabel("Power [dB]")
 
     def set_polarization_model(self, model: QStringListModel) -> None:
         self.pol_combobox.setModel(model)
 
     def reset(self) -> None:
         self.plot.reset_plot()
-        self.min_spinbox.setValue(self.plot.rmin)
-        self.max_spinbox.setValue(self.plot.rmax)
-        self.step_spinbox.setValue(self.plot.rstep)
+        self.min_spinbox.setValue(self.plot.ymin)
+        self.max_spinbox.setValue(self.plot.ymax)
+        self.step_spinbox.setValue(self.plot.ystep)
 
     def _connect_signals(self) -> None:
         self.pol_combobox.currentTextChanged.connect(self._send_controls_state)
-        self.freq_spinbox.editingFinished.connect(self._send_controls_state)
+        self.az_spinbox.valueChanged.connect(self._send_controls_state)
+        self.el_spinbox.valueChanged.connect(self._send_controls_state)
 
         self.min_spinbox.valueChanged.connect(self._on_plot_min_changed)
         self.max_spinbox.valueChanged.connect(self._on_plot_max_changed)
         self.step_spinbox.valueChanged.connect(self._on_plot_step_changed)
 
         self.autoscale_chkbox.stateChanged.connect(self._on_autoscale_toggle_changed)
+        self.autoscale_btn.clicked.connect(self.plot.autoscale_plot)
         self.plot.autoscaled.connect(self._on_autoscaled)
 
     def _send_controls_state(self) -> None:
         log.debug("Controls updated. Sending...")
         pol = self.pol_combobox.currentText()
-        freq = self.freq_spinbox.text()
+        az = self.az_spinbox.value()
+        el = self.el_spinbox.value()
 
-        ctrl = PlotControls(polarization=pol, frequency=freq, elevation=0.0)
+        ctrl = PlotControls(polarization=pol, azimuth=az, elevation=el)
         self.new_data_requested.emit(ctrl)
 
     def _on_plot_min_changed(self, val: int) -> None:
         if val >= self.max_spinbox.value():
             return
 
-        self.plot.rmin = val
+        self.plot.ymin = val
 
     def _on_plot_max_changed(self, val: int) -> None:
         if val <= self.min_spinbox.value():
             return
 
-        self.plot.rmax = val
+        self.plot.ymax = val
 
     def _on_plot_step_changed(self, val: int) -> None:
-        self.plot.rstep = val
+        self.plot.ystep = val
 
     def _on_autoscale_toggle_changed(self, val: bool) -> None:
         self.plot.autoscale = val
@@ -91,15 +90,20 @@ class PolarPlot(PyChamberPlot):
     def rx_updated_data(self, ntwk: skrf.Network) -> None:
         log.debug("Got new data. Updating...")
         pol = self.pol_combobox.currentText()
-        freq = self.freq_spinbox.text()
-        if ntwk.params['polarization'] != pol:
+        az = self.az_spinbox.value()
+        el = self.el_spinbox.value()
+
+        if (
+            ntwk.params['polarization'] != pol
+            or ntwk.params['azimuth'] != az
+            or ntwk.params['elevation'] != el
+        ):
             return
 
-        theta = np.deg2rad(float(ntwk.params['azimuth']))
-        r = ntwk[freq].s_db  # type: ignore
+        freq = ntwk.frequency.f
+        mags = ntwk.s_db
 
-        log.debug(f"Updating polar plot with {theta=} {r=}")
-        self.plot.update_plot(theta, r)
+        self.plot.update_plot(freq, mags)
 
     def _add_widgets(self) -> None:
         layout = QVBoxLayout(self)
@@ -112,13 +116,17 @@ class PolarPlot(PyChamberPlot):
         self.pol_combobox = QComboBox(self)
         hlayout.addWidget(self.pol_combobox)
 
-        freq_label = QLabel("Frequency", self)
-        hlayout.addWidget(freq_label)
+        az_label = QLabel("Azimuth", self)
+        hlayout.addWidget(az_label)
 
-        self.freq_spinbox = FrequencySpinBox(self)
-        self.freq_spinbox.setSizePolicy(size_policy["PREF_PREF"])
-        self.freq_spinbox.setMinimumWidth(100)
-        hlayout.addWidget(self.freq_spinbox)
+        self.az_spinbox = QDoubleSpinBox(self)
+        hlayout.addWidget(self.az_spinbox)
+
+        el_label = QLabel("Elevation", self)
+        hlayout.addWidget(el_label)
+
+        self.el_spinbox = QDoubleSpinBox(self)
+        hlayout.addWidget(self.el_spinbox)
 
         spacer = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
         hlayout.addItem(spacer)
@@ -160,5 +168,5 @@ class PolarPlot(PyChamberPlot):
 
         layout.addLayout(hlayout)
 
-        self.plot = MplPolarWidget(self)
+        self.plot = MplRectWidget(self)
         layout.addWidget(self.plot)
