+++ pychamber/widgets/__init__.py
@@ -2,4 +2,3 @@ from .experiment import ExperimentType, ExperimentWidget
 from .freq_lineedit import FrequencyLineEdit
 from .freq_spin_box import FrequencySpinBox
 from .freq_validator import FrequencyValidator
-from .plots import PlotsWidget
