--- pychamber/widgets/plots/__init__.py
@@ -1,1 +1,0 @@
-from .plots_container import PlotsWidget
