+++ pyproject.toml
@@ -18,7 +18,7 @@ classifiers = [
 packages = [{ include = "pychamber" }, { include = "tests", format = "sdist" }]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<=3.11"
+python = ">=3.10,<3.11"
 
 black = { version = "^22.3.0", optional = true }
 isort = { version = "^5.8.0", optional = true }
